rest/requirements.txt
arrow==1.2.3
body-oc==2.0.0
config-oc==1.0.3
define-oc==1.0.0
record-mysql==1.0.0
record-redis==1.0.0
tools-oc==1.2.3